+++ src/services/sections/Credit.section.js
@@ -20,62 +20,44 @@ export default class CreaditSection {
     }
 
     get(options) {
+        const that = this;
 
-        let that = this;
+        function getCreditCardText(options) {
+            const result = options.isRefund ? that.$translate.getText('RETURNED_IN_CREDIT_FROM') : that.$translate.getText('PAID_IN_CREDIT_FROM');
 
-        function getCreaditFromText(options) {
-
-            let isRefund = options.isRefund;
-            let issuer = options.issuer;
-
-            let result = "";
-
-            if (isRefund) {
-                result += that.$translate.getText('RETURNED_IN_CREDIT_FROM');
-            } else {
-                result += that.$translate.getText('PAID_IN_CREDIT_FROM');
-            }
-
-            result += ` ${ issuer }`;
-            return result;
+            return result + ` ${options.issuer}`;
         }
 
-        let variables = options.variables;
-        let collections = options.collections;
-        let documentInfo = options.documentInfo;
-        let isRefund = options.isRefund;
+        const payment = _.get(options, 'collections.CREDIT_PAYMENTS[0]');
 
-        let payment = collections.CREDIT_PAYMENTS[0];
-
-        let creaditContainer = this.$htmlCreator.createSection({
+        const creditContainer = this.$htmlCreator.createSection({
             id: 'creadit-section',
             classList: ['creadit-section']
         });
 
-        //#region Creadit from
-
-        let elementCreaditCardText = this.$htmlCreator.create({
+        const elementCreaditCardText = this.$htmlCreator.create({
             id: 'creadit-card-text',
             classList: ['total-name'],
-            value: getCreaditFromText({
-                isRefund: isRefund,
+            value: getCreditCardText({
+                isRefund: options.isRefund,
                 issuer: payment.ISSUER
             })
         });
 
-        let classList = ['total-amount'];
-        let negativeClass = this.$utils.isNegative(payment.P_AMOUNT);
+        const classList = ['total-amount'];
+        const negativeClass = this.$utils.isNegative(payment.P_AMOUNT);
+
         if (negativeClass !== "") {
             classList.push(negativeClass);
         }
 
-        let elementCreaditCardValue = this.$htmlCreator.create({
+        const elementCreaditCardValue = this.$htmlCreator.create({
             id: 'creadit-card-value',
             classList: classList,
             value: this.$utils.toFixedSafe(payment.P_AMOUNT || 0, 2) || ''
         });
 
-        let elementCreaditCardContainer = this.$htmlCreator.create({
+        const elementCreaditCardContainer = this.$htmlCreator.create({
             id: 'creadit-card-container',
             classList: ['itemDiv', 'bold'],
             children: [
@@ -84,34 +66,30 @@ export default class CreaditSection {
             ]
         });
 
-        creaditContainer.append(elementCreaditCardContainer);
+        creditContainer.append(elementCreaditCardContainer);
 
-        //#endregion
+        const P_CHANGE = _.get(payment, 'P_CHANGE');
 
-        //#region Creadit change
-
-        let P_CHANGE = _.get(payment, 'P_CHANGE');
         if (P_CHANGE && P_CHANGE !== 0) {
-
-            let elementChangeText = this.$htmlCreator.create({
+            const elementChangeText = this.$htmlCreator.create({
                 id: 'creadit-change-text',
                 classList: ['total-name'],
                 value: this.$translate.getText('CHANGE_TIP')
             });
 
-            let classList = ['total-amount'];
-            let negativeClass = this.$utils.isNegative(payment.P_CHANGE);
+            const classList = ['total-amount'];
+            const negativeClass = this.$utils.isNegative(payment.P_CHANGE);
             if (negativeClass !== "") {
                 classList.push(negativeClass);
             }
 
-            let elementChangeValue = this.$htmlCreator.create({
+            const elementChangeValue = this.$htmlCreator.create({
                 id: 'creadit-change-value',
                 classList: classList,
                 value: this.$utils.toFixedSafe(payment.P_CHANGE || 0, 2) || ''
             });
 
-            let elementChangeContainer = this.$htmlCreator.create({
+            const elementChangeContainer = this.$htmlCreator.create({
                 id: 'change-container',
                 classList: ['itemDiv', 'bold'],
                 children: [
@@ -120,42 +98,35 @@ export default class CreaditSection {
                 ]
             });
 
-            creaditContainer.append(elementChangeContainer);
-
+            creditContainer.append(elementChangeContainer);
         }
 
-        //#endregion
-
-        //#region Creadit change
-
-        let len = _.get(collections, 'CREDIT_PAYMENTS[0].EMV.length') || 0;
+        const len = _.get(payment, 'EMV.length', 0);
         const documentType = _.get(options, 'documentInfo.documentType');
         const hasInstallmentsPayment = !!payment.INSTALLMENTS_COUNT;
 
         if (hasInstallmentsPayment) {
             this.installmentsSection = new InstallmentsSection(this.options, payment, documentInfo);
-            creaditContainer.append(this.installmentsSection.get());
+            creditContainer.append(this.installmentsSection.get());
         }
 
         if (documentType === 'invoice' && len > 0) {
             const elementEmv = this.$emvService.createEmvTemplate(documentType, {
-                variables: variables,
-                collections: collections
+                variables: options.variables,
+                collections: options.collections
             }, this._doc);
 
-            creaditContainer.append(elementEmv);
+            creditContainer.append(elementEmv);
 
         } else if (payment) {
-            let elementCreditTransaction = this.$creditTransaction.get({
+            const elementCreditTransaction = this.$creditTransaction.get({
                 isUS: this.$localization.isUS,
                 data: payment
             });
 
-            creaditContainer.append(elementCreditTransaction);
+            creditContainer.append(elementCreditTransaction);
         }
 
-        //#endregion
-
-        return creaditContainer;
+        return creditContainer;
     }
 }