+++ src/services/sections/Payments.section.js
@@ -4,7 +4,6 @@ import TlogDocsTranslateService from '../../tlog-docs-template/tlogDocsTranslate
 import { InstallmentsSection } from "./Installments";
 
 export default class PaymentSection {
-
     constructor(options) {
         this.$htmlCreator = new HtmlCreator();
         this.$translate = new TlogDocsTranslateService(options);
@@ -13,21 +12,17 @@ export default class PaymentSection {
     }
 
     get(options) {
-        let variables = options.variables;
-        let collections = options.collections;
         let payments = options.payments;
-
         let paymentSection = this.$htmlCreator.createSection({
             id: 'payment-section',
             classList: ['payment-section']
         });
 
-
         payments.forEach(payment => {
             let elementText = this.$htmlCreator.create({
                 id: 'payment-text',
                 classList: ['total-name'],
-                value: payment.name || ''
+                value: payment.CURRENCY_FACE_VALUE ? '' : payment.name || ''
             });
 
             let classList = ['total-amount'];
@@ -42,12 +37,12 @@ export default class PaymentSection {
                 value: this.$utils.toFixedSafe(payment.amount || 0, 2) || ''
             });
 
-            if (payment.CURRENCY_AMOUNT) {
-                const currencyValue = this.$utils.toFixedSafe(payment.CURRENCY_AMOUNT, 2);
+            if (payment.CURRENCY_FACE_VALUE) {
                 const currencySymbol = payment.CURRENCY_SYMBOL;
+                const paymentLabel = this.$translate.getText(`CURRENCY_PAYMENT_LABEL_${currencySymbol}`);
                 const temp = this.$htmlCreator.create({
-                    classList: ['payment-currency'],
-                    value: `(${currencySymbol}${currencyValue})`
+                    classList: ['bold'],
+                    value: paymentLabel
                 });
 
                 elementText.appendChild(temp);
@@ -75,9 +70,34 @@ export default class PaymentSection {
                     elementValue
                 ]
             });
-
-
             paymentSection.append(elementPaymentContainer);
+
+            if (payment.CURRENCY_FACE_VALUE) {
+                let currencyPaymentDetailsText = this.$htmlCreator.create({
+                    id: 'payment-text',
+                    classList: ['total-name'],
+                    value: payment.CURRENCY_FACE_VALUE ? '' : payment.name || ''
+                });
+                const currencyValue = this.$utils.toFixedSafe(payment.CURRENCY_FACE_VALUE, 2);
+                const currencyRate = this.$utils.toFixedSafe(payment.CURRENCY_RATE, 3);
+                const currencyPaymentDetails = this.$htmlCreator.create({
+                    classList: ['bold'],
+                    value: this.$translate.getText(`CURRENCY_PAYMENT_DETAILS_${payment.CURRENCY_SYMBOL}`,
+                        ['currencyAmount', 'currencySymbol', 'currencyRate'],
+                        [currencyValue, payment.CURRENCY_SYMBOL, currencyRate])
+                });
+                currencyPaymentDetailsText.appendChild(currencyPaymentDetails);
+
+                let currencyDetailsContainer = this.$htmlCreator.create({
+                    id: 'payment-container',
+                    classList: classListContainer,
+                    children: [
+                        currencyPaymentDetailsText,
+                    ]
+                });
+
+                paymentSection.append(currencyDetailsContainer);
+            }
 
             if (payment.name.includes('Cash') && payment.P_BONUS_AMOUNT !== undefined) {
                 let elementText = this.$htmlCreator.create({
@@ -100,6 +120,7 @@ export default class PaymentSection {
                         elementValue
                     ]
                 });
+
                 paymentSection.append(elementPaymentContainer);
             }
 
@@ -124,9 +145,9 @@ export default class PaymentSection {
                         elementValue
                     ]
                 });
+
                 paymentSection.append(elementPaymentContainer);
             }
-
 
             if (payment.holderName) {
 
@@ -139,37 +160,46 @@ export default class PaymentSection {
                 paymentSection.append(elementHolderName);
             }
 
-            if (payment.HOTEL_NAME || payment.GUEST_NAME || payment.ROOM_NUMBER || payment.HOTEL_CHECK_NUMBER) {
+            if (payment.HOTEL_NAME || payment.ROOM_NUMBER) {
+                let elementHotelDetails = this.$htmlCreator.create({
+                    id: 'hotel-details',
+                    classList: ['hotel-item', 'hotel-details'],
+                    value: `&nbsp;${payment.HOTEL_NAME}&nbsp;/&nbsp;${payment.ROOM_NUMBER}`
+                });
 
-                if (payment.HOTEL_NAME || payment.ROOM_NUMBER) {
-                    let elementHotelDetails = this.$htmlCreator.create({
-                        id: 'hotel-details',
-                        classList: ['hotel-item', 'hotel-details'],
-                        value: `&nbsp;${payment.HOTEL_NAME}&nbsp;/&nbsp;${payment.ROOM_NUMBER}`
-                    });
-                    paymentSection.append(elementHotelDetails);
-                }
-
-
-                if (payment.GUEST_NAME.replace(/\s/g, '').length > 0) {
-                    let elementGuestName = this.$htmlCreator.create({
-                        id: 'guest-name',
-                        classList: ['hotel-item', 'guest-name'],
-                        value: `&nbsp;${payment.GUEST_NAME}`
-                    });
-
-                    paymentSection.append(elementGuestName);
-                }
+                paymentSection.append(elementHotelDetails);
+            }
 
 
-                if(payment.HOTEL_CHECK_NUMBER) {
-                    let elementHotelCheckNumber = this.$htmlCreator.create({
-                        id: 'hotel-check-number',
-                        classList: ['hotel-item', 'guest-name'],
-                        value: `&nbsp;${this.$translate.getText('HOTEL_CHECK_NUMBER')}&nbsp;${payment.HOTEL_CHECK_NUMBER}`
-                    });
-                    paymentSection.append(elementHotelCheckNumber);
-                }
+            if (payment.GUEST_NAME && payment.GUEST_NAME.replace(/\s/g, '').length > 0) {
+                let elementGuestName = this.$htmlCreator.create({
+                    id: 'guest-name',
+                    classList: ['hotel-item', 'guest-name'],
+                    value: `&nbsp;${payment.GUEST_NAME}`
+                });
+
+                paymentSection.append(elementGuestName);
+            }
+
+            if (payment.HOTEL_CHECK_NUMBER) {
+                let elementHotelCheckNumber = this.$htmlCreator.create({
+                    id: 'hotel-check-number',
+                    classList: ['hotel-item', 'guest-name'],
+                    value: `&nbsp;${this.$translate.getText('HOTEL_CHECK_NUMBER')}&nbsp;${payment.HOTEL_CHECK_NUMBER}`
+                });
+
+                paymentSection.append(elementHotelCheckNumber);
+            }
+
+            if (payment.HOTEL_NAME && payment.CURRENCY_SYMBOL) {
+                const currencyValueText = payment.CURRENCY_SYMBOL + payment.CURRENCY_AMOUNT;
+                const elementHotelRoomChargePayment = this.$htmlCreator.create({
+                    id: 'room-charge-payment',
+                    classList: ['bold'],
+                    value: `&nbsp;${this.$translate.getText('HOTELS_ROOM_CHARGE_CURRENCY', ['value'], [currencyValueText])}`
+                });
+
+                paymentSection.append(elementHotelRoomChargePayment);
             }
 
             if (payment.PROVIDER_TRANS_ID) {
@@ -178,6 +208,7 @@ export default class PaymentSection {
                     classList: ['hotel-item', 'hotel-details'],
                     value: `&nbsp;${this.$translate.getText('PROVIDER_TRANS_ID')} ${payment.PROVIDER_TRANS_ID}&nbsp;`
                 });
+
                 paymentSection.append(elementProviderTransId);
             }